requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0 